Podzial HelicopterController na osobne sktypty
