+++ Assets/Art/CREATE.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: 1261bcdf7c5716545881385851604520
 DefaultImporter:
   externalObjects: {}
   userData:
