+++ Assets/Art/GUZIKI.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: 4e316954baf3f324dabe8ceb9701af77
 DefaultImporter:
   externalObjects: {}
   userData:
