+++ Assets/Art/MAIN.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: c088332ffdf283e418478c43430d1c93
 DefaultImporter:
   externalObjects: {}
   userData:
