+++ Assets/Art/RESTART.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: 1882ee8118bed5e448eaa56e2746d900
 DefaultImporter:
   externalObjects: {}
   userData:
