+++ Assets/Art/RESUME.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: f3502b895b6a6f44d8369a19e72bad41
 DefaultImporter:
   externalObjects: {}
   userData:
