+++ Assets/Art/RETORN.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: 65ac3a9781ad8d94fa8a83e58c48413d
 DefaultImporter:
   externalObjects: {}
   userData:
