+++ Assets/Art/SLOT1.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: 8da9d61eb9c012a40ae72d5a6cb6f6c8
 DefaultImporter:
   externalObjects: {}
   userData:
