+++ Assets/Art/SLOT2.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: ebc82914e3ec3504485330d361d112c7
 DefaultImporter:
   externalObjects: {}
   userData:
