+++ Assets/Art/SLOT3.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 638943a93d6c51f449f00db071ba70ba
+guid: 612f2776acfa1ac4daad1b83b32506e9
 DefaultImporter:
   externalObjects: {}
   userData:
