+++ Assets/Scripts/Player/HelicopterController.cs
@@ -11,18 +11,11 @@ public class HelicopterController : MonoBehaviour
     public float rotationSpeed = 10f;
     public float friction = 10f;
 
-    public TMP_Text timerText;
-    private float timeElapsed;
-
     private Vector3 velocity = Vector3.zero;
     private Vector3 inputDirection = Vector3.zero;
 
     void Update()
     {
-       // timeElapsed += Time.deltaTime;
-        int seconds = Mathf.FloorToInt(timeElapsed % 60);
-       // timerText.text = $"{seconds:000}";
-
         // Input for either keyboard or controller
         float moveX = Input.GetAxis("Horizontal");
         float moveZ = Input.GetAxis("Vertical");
@@ -51,38 +44,6 @@ public class HelicopterController : MonoBehaviour
             Quaternion targetRotation = Quaternion.LookRotation(velocity.normalized);
             transform.rotation = Quaternion.Slerp(transform.rotation, targetRotation, rotationSpeed * Time.deltaTime);
         }
-    }
-
-    private void OnCollisionEnter(Collision collision)
-    {
-        UnityEngine.Debug.Log("Hit: " + collision.gameObject.name);
-        velocity = Vector3.Reflect(velocity, collision.contacts[0].normal) * 0.5f;
-    }
-
-    public Slider healthSlider;
-
-    void UpdateHealth(float currentHealth, float maxHealth)
-    {
-        healthSlider.value = currentHealth / maxHealth;
-    }
-
-    //public Image weaponIcon;
-    public Sprite rocketSprite;
-    public Sprite minigunSprite;
-
-    void SwitchWeapon(int weaponId)
-    {
-        //if (weaponId == 0) weaponIcon.sprite = rocketSprite;
-        //if (weaponId == 1) weaponIcon.sprite = minigunSprite;
-    }
-
-    public TMP_Text scoreText;
-    private int score = 0;
-
-    public void AddScore(int amount)
-    {
-        score += amount;
-        scoreText.text = "Score: " + score;
     }
 
     public TMP_Text objectiveText;
