+++ Assets/Scripts/Player/ScoreCounter.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using TMPro;
+using UnityEngine;
+
+public class ScoreCounter : MonoBehaviour
+{
+    public TMP_Text scoreText;
+    private int score = 0;
+
+    public void AddScore(int amount)
+    {
+        score += amount;
+        scoreText.text = "Score: " + score;
+    }
+}
