+++ Assets/Scripts/Player/ScoreCounter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91385f4c71bc5a3499ad005906f53dbf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
