+++ Assets/Scripts/Player/TimeCounter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a36d6755c5944d04fbbeddab475788f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
