+++ Assets/Scripts/Player/WeaponSwitch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 587d5b8b86aacca42a170f82022ec7c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
